+++ src/main/java/com/htp/service/impl/DoughTypeServiceImpl.java
@@ -1,6 +1,7 @@
 package com.htp.service.impl;
 
 import com.htp.domain.DoughType;
+import com.htp.exception.EntityNotFoundException;
 import com.htp.repository.DoughTypeRepository;
 import com.htp.service.DoughTypeService;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -35,11 +37,20 @@ public class DoughTypeServiceImpl implements DoughTypeService {
     @Transactional
     @Override
     public void delete(Long id) {
-        doughTypeRepository.deleteById(id);
+        if (doughTypeRepository.findById(id).isPresent()) {
+            doughTypeRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(DoughType.class, id);
+        }
     }
 
     @Override
     public DoughType findById(Long id) {
-        return doughTypeRepository.findById(id).get();
+        Optional<DoughType> result = doughTypeRepository.findById(id);
+        if (result.isPresent()) {
+            return result.get();
+        } else {
+            throw new EntityNotFoundException(DoughType.class, id);
+        }
     }
 }
