+++ src/main/java/com/htp/service/impl/PriceServiceImpl.java
@@ -1,6 +1,7 @@
 package com.htp.service.impl;
 
 import com.htp.domain.Price;
+import com.htp.exception.EntityNotFoundException;
 import com.htp.repository.PriceRepository;
 import com.htp.service.PriceService;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -35,12 +37,21 @@ public class PriceServiceImpl implements PriceService {
     @Transactional
     @Override
     public void delete(Long id) {
-        priceRepository.deleteById(id);
+        if (priceRepository.findById(id).isPresent()) {
+            priceRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(Price.class, id);
+        }
     }
 
     @Override
     public Price findById(Long id) {
-        return priceRepository.findById(id).get();
+        Optional<Price> result = priceRepository.findById(id);
+        if (result.isPresent()) {
+            return result.get();
+        } else {
+            throw new EntityNotFoundException(Price.class, id);
+        }
     }
 
     @Override
