+++ src/main/java/com/htp/service/impl/RoleServiceImpl.java
@@ -1,6 +1,7 @@
 package com.htp.service.impl;
 
 import com.htp.domain.Role;
+import com.htp.exception.EntityNotFoundException;
 import com.htp.repository.RoleRepository;
 import com.htp.service.RoleService;
 import lombok.RequiredArgsConstructor;
@@ -8,6 +9,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -35,12 +37,21 @@ public class RoleServiceImpl implements RoleService {
     @Transactional
     @Override
     public void delete(Long id) {
-        roleRepository.deleteById(id);
+        if (roleRepository.findById(id).isPresent()) {
+            roleRepository.deleteById(id);
+        } else {
+            throw new EntityNotFoundException(Role.class, id);
+        }
     }
 
     @Override
     public Role findById(Long id) {
-        return roleRepository.findById(id).get();
+        Optional<Role> result = roleRepository.findById(id);
+        if (result.isPresent()) {
+            return result.get();
+        } else {
+            throw new EntityNotFoundException(Role.class, id);
+        }
     }
 
     @Override
