+++ src/main/java/com/htp/service/impl/UserServiceImpl.java
@@ -2,6 +2,7 @@ package com.htp.service.impl;
 
 
 import com.htp.domain.User;
+import com.htp.exception.EntityNotFoundException;
 import com.htp.repository.UserRepository;
 import com.htp.service.UserService;
 import lombok.RequiredArgsConstructor;
@@ -9,6 +10,7 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 @RequiredArgsConstructor
@@ -36,12 +38,21 @@ public class UserServiceImpl implements UserService {
     @Transactional
     @Override
     public void delete(Long id) {
-        userRepository.delete(id);
+        if(userRepository.findById(id).isPresent()) {
+            userRepository.delete(id);
+        } else {
+            throw new EntityNotFoundException(User.class, id);
+        }
     }
 
     @Override
     public User findById(Long id) {
-        return userRepository.findById(id).get();
+        Optional<User> result = userRepository.findById(id);
+        if(result.isPresent()) {
+            return result.get();
+        } else {
+            throw new EntityNotFoundException(User.class, id);
+        }
     }
 
     @Override
